backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.2.2
torchvision==0.17.2
timm==0.9.10
pillow>=10.0.0
numpy>=1.24.3
python-multipart==0.0.6 